backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.35
pydantic-settings==2.5.2
google-generativeai==0.8.3
Pillow==12.2.0
